enable download and zoom actions. skip freeze action for now.
This brings the original download and zoom actions into a more visible
and discoverable interaction. We defer the work of creating a freeze
action for now in the interest of getting this much released.

Freeze and thaw behavior will need some care to ensure it is only
available when editing and only for dynamically computed images.